refactor: change name to some method

+++ app/src/services/user.py
@@ -1,12 +1,17 @@
+from flask import abort
 from src.models import UserRecord, User
 from src.firebase import web_sdk
 
 
+
 class UserService:
 
-    def get_user(self, uid):
+    def get(self, uid):
         user_record = UserRecord.get_user(uid, app=web_sdk)
         user = User.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         return {
             'uid': user_record.uid,
@@ -14,13 +19,12 @@ class UserService:
             'b': user
         }
 
-    def list_user(self, page=1):
-        per_pages = 10
-        users = User.query.paginate(page, per_pages, error_out=False)
+    def list(self, page=1, per_pages=10):
+        users = User.query.paginate(page, per_pages or 10, error_out=False)
         
         return users
 
-    def create_user(self, body):
+    def create(self, body):
         try:
             user_record = UserRecord(
                     email=body['email'],
@@ -50,9 +54,12 @@ class UserService:
         except KeyError as ex:
             abort(400, description='BadRequest', response=str(ex))
 
-    def update_user(self, uid, body):
+    def update(self, uid, body):
         user_record = UserRecord.get_user(uid, app=web_sdk) 
         user = User.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize(body)
         user_record.update_user()
@@ -69,9 +76,12 @@ class UserService:
             'b': user
         }
 
-    def update_field_user(self, uid, body):
+    def update_field(self, uid, body):
         user_record = UserRecord.get_user(uid, app=web_sdk) 
         user = User.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize(body)
         user_record.update_user()
@@ -88,9 +98,12 @@ class UserService:
             'b': user
         }
 
-    def disabled_user(self, uid):
+    def disabled(self, uid):
         user_record = UserRecord.get_user(uid, app=web_sdk)
         user = User.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.serialize({ 'disabled': not user_record.disabled })
         user_record.update_user()
@@ -105,9 +118,12 @@ class UserService:
         }
 
 
-    def delete_user(self, uid):
+    def delete(self, uid):
         user_record = UserRecord.get_user(uid, app=web_sdk)
         user = User.query.filter_by(uid=user_record.uid).first()
+
+        if not user_record or not user:
+            abort(404, description='NotFound', response='not_found')
 
         user_record.delete_user()
         user.delete()
